+++ src/main/java/dev/dburns/budget_tracker/model/User.java
@@ -10,7 +10,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     private String name;
 
@@ -21,13 +21,13 @@ public class User {
     public User() {
     }
 
-    public User(int id, String name, String email) {
+    public User(long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
-    public int getId()
+    public long getId()
     {
         return id;
     }
@@ -40,7 +40,7 @@ public class User {
         return email;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
